Add 64-bit support?

+++ src/utils.h
@@ -31,10 +31,10 @@ THE SOFTWARE.
 #include <QString>
 #include <math.h>
 
-// valid for as long as DF stays 32bit
-typedef quint32 VIRTADDR;
-typedef quint32 USIZE;
-typedef qint32 SSIZE;
+// shit
+typedef quint64 VIRTADDR;
+typedef quint64 USIZE;
+typedef qint64 SSIZE;
 typedef quint8 BYTE;
 typedef quint16 WORD;
